+++ NEWS.md
@@ -1,3 +1,4 @@
 # regtools (development version)
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added `plot_map()` function to create Norwegian choropleth municipality and country maps